Updating medium fishing and int random

+++ SA2B_Archipelago/SA2B_Archipelago/Items/Minigames/MinigameUtilities.h
@@ -13,4 +13,4 @@ std::default_random_engine& RNG();
 
 float RandomFloat(float min, float max);
 
-float RandomInt(int minInc, int maxExcl);
+int RandomInt(int minInc, int maxExcl);
